Feat: seven the folder as src of test
