+++ src/main/java/brawl/example/brawl_tracker/ProjectBrawlApiSheetsApplication.java
@@ -1,0 +1,15 @@
+package brawl.example.project_brawl_api_sheets;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class ProjectBrawlApiSheetsApplication {
+
+	public static void main(String[] Args) {
+		SpringApplication.run(ProjectBrawlApiSheetsApplication.class, Args);
+	}
+
+}
