+++ src/main/java/brawl/example/brawl_tracker/brawl_core/config/AppConfig.java
@@ -1,0 +1,14 @@
+package brawl.example.brawl_tracker.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
