+++ src/main/java/brawl/example/brawl_tracker/brawl_core/model/PlayerValidModel.java
@@ -1,0 +1,12 @@
+package brawl.example.brawl_tracker.model;
+
+import lombok.*;
+
+@Data
+public class PlayerValidModel {
+
+    @NonNull
+    private String name;
+    private Integer trophies;
+    private boolean isQualifiedFromChampionshipChallenge;
+}
