+++ src/main/java/brawl/example/brawl_tracker/brawl_core/model/TeamPlayerTag.java
@@ -1,0 +1,10 @@
+package brawl.example.brawl_tracker.model;
+
+import java.util.List;
+
+public interface TeamPlayerTag {
+
+    String getTeamName();
+    List<String> getTags();
+
+}
