+++ src/main/java/brawl/example/brawl_tracker/brawl_core/util/OpenURL.java
@@ -1,0 +1,12 @@
+package brawl.example.brawl_tracker.util;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+public class OpenURL {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
