--- src/main/java/brawl/example/project_brawl_api_sheets/ProjectBrawlApiSheetsApplication.java
@@ -1,13 +1,0 @@
-package brawl.example.project_brawl_api_sheets;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProjectBrawlApiSheetsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProjectBrawlApiSheetsApplication.class, args);
-	}
-
-}
