--- src/main/java/brawl/example/project_brawl_api_sheets/config/BotDiscordAPIConfig.java
@@ -1,37 +1,0 @@
-package brawl.example.project_brawl_api_sheets.config;
-
-import brawl.example.project_brawl_api_sheets.controller.BotController;
-import jakarta.annotation.PostConstruct;
-import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.JDABuilder;
-import net.dv8tion.jda.api.interactions.commands.build.Commands;
-import net.dv8tion.jda.api.requests.GatewayIntent;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class BotDiscordAPIConfig {
-
-    private final BotController botListener;
-
-    @Value("${bot.discord.token}")
-    private String token;
-
-    public BotDiscordAPIConfig(BotController botListener) {
-        this.botListener = botListener;
-    }
-
-    @PostConstruct
-    public void startBot() throws Exception {
-        JDA jda = JDABuilder.createDefault(token, GatewayIntent.GUILD_MESSAGES,
-                        GatewayIntent.MESSAGE_CONTENT)
-                .addEventListeners(botListener)
-                .build();
-
-        jda.updateCommands()
-                .addCommands(
-                        Commands.slash("infoteams", "Contribute registering the teams")
-                )
-                .queue();
-    }
-}
