--- src/main/java/brawl/example/project_brawl_api_sheets/model/TeamMODEL.java
@@ -1,13 +1,0 @@
-package brawl.example.project_brawl_api_sheets.model;
-import lombok.Data;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.ModelAttribute;
-
-import java.util.*;
-
-@Data
-@Component
-public class TeamMODEL {
-    private String teamName;
-    private List<String> playersTags;
-}
