--- src/main/java/brawl/example/project_brawl_api_sheets/model/TeamPlayerTag.java
@@ -1,12 +1,0 @@
-package brawl.example.project_brawl_api_sheets.model;
-
-import brawl.example.project_brawl_api_sheets.entity.PlayerTagEntity;
-
-import java.util.List;
-
-public interface TeamPlayerTag {
-
-    String getTeamName();
-    List<String> getTags();
-
-}
