--- src/main/java/brawl/example/project_brawl_api_sheets/util/OpenURL.java
@@ -1,12 +1,0 @@
-package brawl.example.project_brawl_api_sheets.util;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
-
-public class OpenURL {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
