--- src/test/java/brawl/example/project_brawl_api_sheets/ProjectBrawlApiSheetsApplicationTests.java
@@ -1,13 +1,0 @@
-package brawl.example.project_brawl_api_sheets;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjectBrawlApiSheetsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
